+++ gatsby-ssr.tsx
@@ -1,5 +1,6 @@
-import { CartProvider, SessionProvider, UIProvider } from '@faststore/sdk'
+import { CartProvider, SessionProvider } from '@faststore/sdk'
 import { validateSession } from 'src/sdk/session/validate'
+import UIProvider from 'src/sdk/ui/Provider'
 import type { ReactNode } from 'react'
 import type { GatsbySSR } from 'gatsby'
 
@@ -9,19 +10,13 @@ import AnalyticsHandler from './src/sdk/analytics'
 import { validateCart } from './src/sdk/cart/validate'
 import ErrorBoundary from './src/sdk/error/ErrorBoundary'
 import TestProvider from './src/sdk/tests'
-import { uiActions, uiEffects, uiInitialState } from './src/sdk/ui'
-import { ModalProvider } from './src/sdk/ui/modal'
 import storeConfig from './store.config'
 
 export const wrapRootElement: GatsbySSR['wrapRootElement'] = ({ element }) => (
   <ErrorBoundary>
     <AnalyticsHandler />
     <TestProvider>
-      <UIProvider
-        initialState={uiInitialState}
-        actions={uiActions}
-        effects={uiEffects}
-      >
+      <UIProvider>
         <SessionProvider
           initialState={{
             channel: storeConfig.channel,
@@ -30,7 +25,7 @@ export const wrapRootElement: GatsbySSR['wrapRootElement'] = ({ element }) => (
           onValidateSession={validateSession}
         >
           <CartProvider mode="optimistic" onValidateCart={validateCart}>
-            <ModalProvider>{element}</ModalProvider>
+            {element}
           </CartProvider>
         </SessionProvider>
       </UIProvider>