+++ src/components/regionalization/RegionalizationInput/RegionalizationInput.tsx
@@ -2,13 +2,15 @@ import { useSession } from '@faststore/sdk'
 import { useRef, useState } from 'react'
 import InputText from 'src/components/ui/InputText'
 import { validateSession } from 'src/sdk/session/validate'
-import { useModal } from 'src/sdk/ui/modal/Provider'
 
-export default function RegionalizationInput() {
+interface Props {
+  closeModal: () => void
+}
+
+function RegionInput({ closeModal }: Props) {
   const inputRef = useRef<HTMLInputElement>(null)
   const { setSession, isValidating, ...session } = useSession()
   const [errorMessage, setErrorMessage] = useState<string>('')
-  const { onModalClose } = useModal()
 
   const handleSubmit = async () => {
     const value = inputRef.current?.value
@@ -29,7 +31,7 @@ export default function RegionalizationInput() {
         setSession(newSession)
       }
 
-      onModalClose()
+      closeModal()
     } catch (error) {
       setErrorMessage('You entered an invalid Zip Code')
     }
@@ -48,3 +50,5 @@ export default function RegionalizationInput() {
     </div>
   )
 }
+
+export default RegionInput